requirements.txt
Fabric==1.8.2
Flask==0.10.1
Flask-SQLAlchemy==1.0
Jinja2==2.7.2
MarkupSafe==0.18
SQLAlchemy==0.9.3
Werkzeug==0.9.4
coverage==3.7.1
ecdsa==0.10
itsdangerous==0.23
paramiko==1.12.2
pycrypto==2.6.1
wsgiref==0.1.2
Flask-WTF==0.9.4
googlemaps==2.2
factual-api==1.6.0
python-dateutil==1.5
waitress==0.8.9